Possibilitado dar nota para as entregas

+++ lib/models/delivery_model.dart
@@ -7,7 +7,7 @@ import 'dart:convert';
 class DeliveryModel {
     int id;
     DateTime deliveryDate;
-    int grade;
+    double grade;
     DateTime gradeGivenDate;
     int assignmentId;
